refactor: extract the workflow category accordion into a shared component (#1607)
